requirements.txt
selenium
flask
webdriver-manager
flask_cors
requests
boto3
beautifulsoup4
python-dotenv
